renamed route files (and update Manifest & __init__)
